Add tests for OpenGL Surface backing stores
All the EGL context & config details where previously hidden as
implementation details in TestGLSurface. It was necessary to expose that
and make it explicit, so you can create multiple surfaces that share a context.
That explicit EGL context & config is contained in class TestEGLContext.

All opengl tests now have a single TestEGLContext (see EmbedderTestContext->egl_context_)
that's used by default when creating the GL Backing Store producer and the
root GL surface that's used for all the OpenGL Renderer callbacks.

The GL Backing Store producer will use that EGL context to create any
GL surface backing stores, when it is configured to do so. Furthermore,
for creating such backing stores, TestGLSurfaces are kinda heavyweight,
so a new TestGLOnscreenOnlySurface was introduced, which is a lighter
variant of that.

The snapshotting mechanism in embedder_test_compositor_gl.cc had to be
modified a bit, since snapshotting GL surfaces is a bit more involved.
(Surface needs to be made current)